backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background Tasks
apscheduler==3.10.4

# HTTP Client (for Google APIs)
httpx==0.25.2

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Image Processing
Pillow==10.1.0
cloudinary==1.36.0

# Environment Variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.8

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1

